Fix recipe repository with call from the dev branch
